extend actions for head of department

+++ DigitNow.Microservice.DocumentManagement/DigitNow.Domain.DocumentManagement/Business/Common/Factories/WorkflowHistoryLogFactory.cs
@@ -19,7 +19,7 @@ namespace DigitNow.Domain.DocumentManagement.Business.Common.Factories
                 Remarks = remarks,
                 OpinionRequestedUntil = opinionRequestedUntil,
                 Resolution = resolution,
-                DestinationDepartmentId = document.DestinationDepartmentId,
+                DestinationDepartmentId = (int)document.DestinationDepartmentId,
             };
     }
 }
